adding provisions for rendering generated article fixing margin issue for bean attributes

+++ app/web/renderer.py
@@ -267,6 +267,7 @@ def render_page_banner(context: Context):
     elif context.page_type == K_CATEGORIES: banner_text = f"🏷️ {context.page}"
     elif context.page_type == K_REGIONS: banner_text = f"📍 {context.page}"
     elif context.page_type == K_RELATED: banner_text = f"🗞️ {context.page.title}"
+    elif context.page_type == GENERATED: banner_text = context.page.title
     else: banner_text = context.page
 
     with render_banner(banner_text) as view:
@@ -305,6 +306,8 @@ def render_similar_pages(context: Context):
         with ui.row().classes("w-full gap-1"):
             render_page_names(related_baristas)
 
+render_grid = lambda: ui.grid().classes(BEANS_GRID_CLASSES)
+
 def render_beans(context: Context, load_beans: Callable, container: ui.element = None, item_classes: str = STRETCH_FIT):
     async def render():
         beans = await run.io_bound(load_beans)
@@ -313,7 +316,7 @@ def render_beans(context: Context, load_beans: Callable, container: ui.element =
             if beans: [render_bean_with_related(context, bean).classes(item_classes) for bean in beans] 
             else: ui.label(NOTHING_FOUND).classes("w-full text-center") 
 
-    container = container or ui.column(align_items="stretch")
+    container = container or render_grid()
     with container:
         render_skeleton_beans(3)
     background_tasks.create_lazy(render(), name=f"beans-{now()}")
@@ -325,6 +328,7 @@ def render_beans_as_extendable_list(context: Context, load_beans: Callable):
     def current_page():
         nonlocal current_start
         beans = load_beans(current_start, config.filters.page.max_beans+1) 
+        if not beans: return
         current_start += config.filters.page.max_beans # moving the cursor
         if len(beans) <= config.filters.page.max_beans:
             more_btn.delete()
@@ -337,7 +341,7 @@ def render_beans_as_extendable_list(context: Context, load_beans: Callable):
                 [render_bean_with_related(context, bean).classes(STRETCH_FIT) for bean in beans[:config.filters.page.max_beans]]
 
     with ui.column() as view:
-        with ui.grid().classes(BEANS_GRID_CLASSES) as beans_panel:
+        with render_grid() as beans_panel:
             render_beans(context, current_page, beans_panel, STRETCH_FIT)
         more_btn = ui.button("More Stories", on_click=next_page).props("rounded no-caps icon-right=chevron_right")
     return view
@@ -407,7 +411,9 @@ def render_expandable_bean(context: Context, bean: Bean, expanded: bool = False,
 def render_whole_bean(context: Context, bean: Bean):
     with ui.column(align_items="stretch") as view:
         render_bean_header(context, bean).classes(add="q-mb-sm")
-        render_bean_body(context, bean)
+        # render_bean_body(context, bean)
+        ui.markdown(bean.content or bean.summary)
+
     return view 
 
 def render_bean(context, bean, expanded: bool = False, on_expanded = None):
@@ -420,58 +426,64 @@ def render_bean_header(context: Context, bean: Bean):
             ui.image(bean.image_url).classes(IMAGE_DIMENSIONS)
         with ui.element().classes("w-full"):
             ui.label(bean.title).classes("bean-title")    
-            render_bean_source_attributes(context, bean).classes("text-caption truncate")
-            render_bean_stats(context, bean).classes("text-caption truncate") 
+            
+            render_bean_source_attributes(context, bean).classes("truncate my-1")
+            render_bean_classification(context, bean).classes("truncate")
+            render_bean_stats(context, bean).classes("truncate my-1") 
     return view
 
 render_link = lambda prefix, title, target: ui.markdown((f"{prefix} " if prefix else "")+f"[{title}]({target})")
-render_navigate_button = lambda title, target: ui.button(title, color="transparent", on_click=lambda: navigate_to(target)).props("no-caps size=sm padding=none flat")
+render_bean_attribute = lambda title, target = None: ui.chip(title, color="transparent", on_click=(lambda: navigate_to(target)) if target else None).props("dense flat").classes("m-0 p-0 text-caption ellipsis")
 
 def render_bean_source_attributes(context: Context, bean: Bean):
     with ui.row(align_items="center", wrap=False).classes("w-full gap-3") as view:
-        ui.label(naturalday(bean.created))
-        render_bean_source(context, bean)
-        if bean.author: ui.label(f"✍️ {bean.author}")
+        render_bean_attribute(naturalday(bean.created))
+        with render_bean_attribute('', create_page_target("/sources", bean.source)) as view:        
+            ui.icon("img:"+ favicon(bean)).classes("mr-1")
+            ui.label(bean.site_name or bean.source)
+        if bean.author: render_bean_attribute(f"✍️ {bean.author}")
+    return view
+
+# def render_bean_source(context: Context, bean: Bean):
+#     with render_bean_attribute('', create_page_target("/sources", bean.source)) as view:        
+#         ui.icon("img:"+ favicon(bean)).classes("mr-1")
+#         ui.label(bean.site_name or bean.source)
+#     return view
+    # with ui.row(wrap=False, align_items="center").classes("gap-1") as view:        
+    #     ui.icon("img:"+ favicon(bean), size="xs")
+    #     render_bean_attribute(bean.site_name or bean.source, create_page_target("/sources", bean.source))
+    # return view
+
+def render_bean_classification(context: Context, bean: Bean): 
+    with ui.row(align_items="center", wrap=False).classes("gap-3") as view:    
+        if bean.categories: render_bean_attribute(f"🏷️ {bean.categories[0]}", create_page_target('/categories', bean.categories[0]))           
+        if bean.regions: render_bean_attribute(f"📍 {bean.regions[0]}", create_page_target('/regions', bean.regions[0]))
     return view
 
 def render_bean_stats(context: Context, bean: Bean): 
-    with ui.row(align_items="center", wrap=False).classes("w-full gap-3") as view:    
-        # if bean.categories: render_navigate_button(f"🏷️ {bean.categories[0]}", create_page_target('/categories', bean.categories[0]))           
-        # if bean.regions: render_navigate_button(f"📍 {bean.regions[0]}", create_page_target('/regions', bean.regions[0]))
-        if bean.categories: render_link("🏷️", bean.categories[0], create_page_target('/categories', bean.categories[0]))           
-        if bean.regions: render_link("📍", bean.regions[0], create_page_target('/regions', bean.regions[0]))
-        if bean.comments: ui.label(f"💬 {humanize.intword(bean.comments)}").tooltip(f"{bean.comments} comments across various social media sources")
-        if bean.likes: ui.label(f"👍 {humanize.intword(bean.likes)}").tooltip(f"{bean.likes} likes across various social media sources")
-        if bean.shares and bean.shares > 1: ui.label(f"🔗 {humanize.intword(bean.shares)}").tooltip(f"{bean.shares} shares across various social media sources") # another option 🗞️
-        if bean.related: render_link("🗞️", bean.related, create_navigation_target('/related', url=bean.url)).tooltip(f"{bean.related} related article(s)")
-    return view
-
-def render_bean_category_and_region(context: Context, bean: Bean):
-    with ui.row(align_items="center", wrap=False).classes("w-full gap-2") as view:
-        if bean.categories: render_navigate_button(f"🏷️ {bean.categories[0]}", create_page_target('/categories', bean.categories[0]))           
-        if bean.regions: render_navigate_button(f"📍 {bean.regions[0]}", create_page_target('/regions', bean.regions[0]))
+    with ui.row(align_items="center", wrap=False).classes("gap-3") as view:    
+        if bean.comments: render_bean_attribute(f"💬 {humanize.intword(bean.comments)}").tooltip(f"{bean.comments} comments across various social media sources")
+        if bean.likes: render_bean_attribute(f"👍 {humanize.intword(bean.likes)}").tooltip(f"{bean.likes} likes across various social media sources")
+        if bean.shares and bean.shares > 1: render_bean_attribute(f"🔗 {humanize.intword(bean.shares)}").tooltip(f"{bean.shares} shares across various social media sources") # another option 🗞️
+        if bean.related: render_bean_attribute(f"🗞️ {bean.related}", create_navigation_target('/related', url=bean.url)).tooltip(f"{bean.related} related article(s)")
     return view
 
 def render_bean_body(context: Context, bean: Bean):
     with ui.column(align_items="stretch").classes("w-full m-0 p-0") as view:
         if bean.entities: render_bean_entities(context, bean)
         if bean.summary: ui.markdown(bean.summary).classes("bean-body truncate-multiline")
+        ui.link("Read More ...", bean.url if bean.kind != GENERATED else create_page_target("/articles", bean.url), new_tab=True).on("click", lambda : context.log("opened", url=bean.url))
         render_bean_actions(context, bean).classes(STRETCH_FIT)
     return view
 
 def render_bean_entities(context: Context, bean: Bean):
     make_tag = lambda tag: ui.link(tag, target=create_page_target("/entities", tag)).classes("tag q-mr-md").style("color: secondary; text-decoration: none;")
+    # make_tag = lambda tag: render_bean_attribute(tag, create_page_target("/entities", tag))
     with ui.row(wrap=True, align_items="baseline").classes("w-full gap-0 m-0 p-0 text-caption") as view:
         list(map(
             make_tag, 
             random.sample(bean.entities, min(config.filters.bean.max_entities, len(bean.entities)))
         ))
-    return view
-
-def render_bean_source(context: Context, bean: Bean):
-    with ui.row(wrap=False, align_items="center").classes("gap-2") as view:        
-        ui.icon("img:"+ favicon(bean))
-        ui.link(bean.site_name or bean.source, bean.url, new_tab=True).classes("truncate").on("click", lambda : context.log("opened", url=bean.url))
     return view
 
 def render_bean_actions(context: Context, bean: Bean): 
@@ -503,7 +515,7 @@ def render_bean_actions(context: Context, bean: Bean):
             #     .set_enabled(context.is_registered)
 
         with ui.button_group().props(ACTION_BUTTON_PROPS).classes("p-0 m-0"):
-            ui.button(icon="rss_feed", color="secondary", on_click=lambda: navigate_to_source(bean.source)).props(ACTION_BUTTON_PROPS).tooltip("More from this channel")
+            # ui.button(icon="rss_feed", color="secondary", on_click=lambda: navigate_to_source(bean.source)).props(ACTION_BUTTON_PROPS).tooltip("More from this channel")
             ui.button(icon="search", color="secondary", on_click=lambda: navigate_to_search(q=bean.url)).props(ACTION_BUTTON_PROPS).tooltip("More like this")
             with ui.button(icon="share", color="secondary").props(ACTION_BUTTON_PROPS):
                 with ui.menu().props("auto-close"):
